midx: test 64-bit offsets
The MIDX feature needs a test that verifies large offsets are written
correctly. To trick git into thinking there are large offsets without
writing 4GB of data, use the --index-version argument to 'git pack-objects'
and then mutate an offset to be large. This causes the IDX file to be
incorrect, but we can verify that the MIDX file notices this offset and
creates a large offset chunk.

In a later change, we will add a 'verify' subcommand to the MIDX so we
can verify that the offset values are copied correctly from the IDX files.

Since the MIDX writes do not actually open the pack-files to test the
offsets, we can make this change safely. We do need to delete the created
IDX and MIDX files so we can avoid breaking tests that may be added after
this one.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>